backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
google-generativeai==0.5.2
python-multipart==0.0.9
python-dotenv==1.0.1
